pyproject.toml
[project]
name = "buurts-devhub"
version = "0.1.0"
description = "Vendor-free development orchestration platform — manages BORIS and other project nodes"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "B", "UP"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.setuptools.packages.find]
include = ["devhub*"]
